passes all tests with removed dependency
