+++ package.json
@@ -33,7 +33,6 @@
     "simple-lru-cache": false,
     "ws": false,
     "fs": false,
-    "xmlhttprequest-ts": false,
     "async-disk-cache": false,
     "distrib/es2015/external/ocsp/ocsp": false,
     "distrib/lib/external/ocsp/ocsp": false,
@@ -106,13 +105,11 @@
     "asn1.js-rfc5280": "^3.0.0",
     "async-disk-cache": "^2.1.0",
     "bent": "^7.3.12",
-    "form-data": "^4.0.0",
     "https-proxy-agent": "^4.0.0",
     "simple-lru-cache": "0.0.2",
     "url-parse": "^1.4.7",
     "uuid": "^3.3.3",
-    "ws": "^7.3.1",
-    "xmlhttprequest-ts": "^1.0.1"
+    "ws": "^7.3.1"
   },
   "resolutions": {
     "extend": "3.0.2"
